Bug caught => xmm use in interrupt handlers breaking floating point code
